refactoring(ResourceExtractor) : deplace le code concernant la collecte des metaux
